+++ src/Aspire.Hosting.AWS/Lambda/SQSEventSourceExtensions.cs
@@ -24,7 +24,7 @@ public static class SQSEventSourceExtensions
 
     /// <summary>
     /// Add an SQS event source to a Lambda function. This feature emulates adding an SQS event source to a Lambda function when deployed to AWS. 
-    /// A separate sub resource will be added to the .NET Aspire application that polls the SQS queue. As messages 
+    /// A separate sub resource will be added to the Aspire application that polls the SQS queue. As messages 
     /// are received from the queue the Lambda function will be invoked with the messages.
     /// </summary>
     /// <param name="lambdaFunction">The Lambda function to add the event source to.</param>
@@ -32,6 +32,7 @@ public static class SQSEventSourceExtensions
     /// <param name="options">Optional configuration for the event source.</param>
     /// <returns></returns>
     /// <exception cref="InvalidOperationException"></exception>
+    [AspireExport("withSQSEventSource")]
     public static IResourceBuilder<LambdaProjectResource> WithSQSEventSource(this IResourceBuilder<LambdaProjectResource> lambdaFunction, string queueUrl, SQSEventSourceOptions? options = null)
     {
         return WithSQSEventSource(lambdaFunction, () => ValueTask.FromResult(queueUrl), options, queueName: null);
@@ -39,7 +40,7 @@ public static class SQSEventSourceExtensions
 
     /// <summary>
     /// Add an SQS event source to a Lambda function. This feature emulates adding an SQS event source to a Lambda function when deployed to AWS. 
-    /// A separate sub resource will be added to the .NET Aspire application that polls the SQS queue. As messages 
+    /// A separate sub resource will be added to the Aspire application that polls the SQS queue. As messages 
     /// are received from the queue the Lambda function will be invoked with the messages.
     /// </summary>
     /// <param name="lambdaFunction">The Lambda function to add the event source to.</param>
@@ -71,7 +72,7 @@ public static class SQSEventSourceExtensions
 
     /// <summary>
     /// Add an SQS event source to a Lambda function. This feature emulates adding an SQS event source to a Lambda function when deployed to AWS. 
-    /// A separate sub resource will be added to the .NET Aspire application that polls the SQS queue. As messages 
+    /// A separate sub resource will be added to the Aspire application that polls the SQS queue. As messages 
     /// are received from the queue the Lambda function will be invoked with the messages.
     /// </summary>
     /// <param name="lambdaFunction">The Lambda function to add the event source to.</param>